[tests/refactor] Created folder src.enum, added unit-test for exceptions
