+++ examples/example.py
@@ -1,14 +1,15 @@
 from src.edocx import EDocx
-from src.utils import Color, UnderlineStyle, FontStyle
+from src.enum.color import Color
+from src.enum.font_style import FontStyle
+from src.enum.underline_style import UnderlineStyle
 
 if __name__ == '__main__':
     doc = EDocx('example.docx')  # Open example.docx
-
     doc.autosave = True  # Enable autosave changes
 
     attrib = doc.all_para_attributes()  # Get all paragraph attributes of a file
 
-    paraId = attrib[0].values()[0]  # Get paraId of first paragraph
+    paraId = attrib[1].values()[0]  # Get paraId of first paragraph
 
     # Add comment to first paragraph with text 'Example comment' by 'Manager' and print result
     doc.add_comment_by_id(paraId, 'Example comment', 'Manager')
@@ -17,7 +18,7 @@ if __name__ == '__main__':
     doc.edit_style_by_id(
         paraId=paraId,
         size=14,
-        color=Color.BLUE,
+        color=Color.PINK,
         italic=False,
         bold=True,
         underline=UnderlineStyle.WAVY_DOUBLE,
