+++ src/enum/font_style.py
@@ -1,0 +1,18 @@
+
+class FontStyle:
+    """
+    Class with font enum
+    """
+    ARIAL = 'Arial'
+    CALIBRI = 'Calibri'
+    TIMES_NEW_ROMAN = 'Times New Roman'
+    COURIER_NEW = 'Courier New'
+    VERDANA = 'Verdana'
+    GEORGIA = 'Georgia'
+    HELVETICA = 'Helvetica'
+    TAHOMA = 'Tahoma'
+    TREBUCHET_MS = 'Trebuchet MS'
+    IMPACT = 'Impact'
+    COMIC_SANS_MS = 'Comic Sans MS'
+    SYMBOL = 'Symbol'
+    WINGDINGS = 'Wingdings'
