+++ src/enum/underline_style.py
@@ -1,0 +1,58 @@
+
+class UnderlineStyle:
+    """
+    Class with enum for underline
+    """
+    NONE = 0
+    """None underline."""
+
+    SINGLE = 1
+    """A single line underline."""
+
+    WORDS = 2
+    """Underline individual words only."""
+
+    DOUBLE = 3
+    """A double line underline."""
+
+    DOTTED = 4
+    """Dotted underline."""
+
+    THICK = 6
+    """A thick line underline."""
+
+    DASH = 7
+    """Dashed underline."""
+
+    DOT_DASH = 9
+    """Alternating dots and dashes underline."""
+
+    DOT_DOT_DASH = 10
+    """An alternating dot-dot-dash pattern underline."""
+
+    WAVY = 11
+    """Wavy underline."""
+
+    DOTTED_HEAVY = 20
+    """Heavy dots underline."""
+
+    DASH_HEAVY = 23
+    """Heavy dashes underline."""
+
+    DOT_DASH_HEAVY = 25
+    """Alternating heavy dots and heavy dashes underline."""
+
+    DOT_DOT_DASH_HEAVY = 26
+    """An alternating heavy dot-dot-dash pattern underline."""
+
+    WAVY_HEAVY = 27
+    """Heavy wavy underline."""
+
+    DASH_LONG = 39
+    """Long dashes underline."""
+
+    WAVY_DOUBLE = 43
+    """Double wavy underline."""
+
+    DASH_LONG_HEAVY = 55
+    """Long heavy dashes underline."""
